- add todo screen logic. - add complete task screen logic.
