+++ lib/model/task_model.dart
@@ -1,10 +1,12 @@
 class TaskModel {
+  final int id;
   final String taskName;
   final String taskDescription;
   final bool isHighPriority;
   bool isDone;
 
   TaskModel({
+    required this.id,
     required this.taskName,
     required this.taskDescription,
     required this.isHighPriority,
@@ -13,6 +15,7 @@ class TaskModel {
 
   factory TaskModel.fromJson(Map<String, dynamic> json) {
     return TaskModel(
+      id: json["id"],
       taskName: json["taskName"],
       taskDescription: json["taskDescription"],
       isHighPriority: json["isHighPriority"],
@@ -22,6 +25,7 @@ class TaskModel {
 
   Map<String, dynamic> toJson() {
     return {
+      "id": id,
       "taskName": taskName,
       "taskDescription": taskDescription,
       "isHighPriority": isHighPriority,
