--- lib/widget/task_completed_screen.dart
@@ -1,9 +1,0 @@
-import 'package:flutter/material.dart';
-
-class TaskCompletedScreen extends StatelessWidget {
-  const TaskCompletedScreen({super.key});
-
-  Widget build(BuildContext context) {
-    return Center(child: Container(child: Text("Task Completed Screen")));
-  }
-}
