--- lib/widget/todo_screen.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class TodoScreen extends StatelessWidget {
-  const TodoScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(child: Container(child: Text("To Do Screen")));
-  }
-}
